+++ Color.h
@@ -9,9 +9,9 @@
 
 class Color {
 public:
-    int r, g, b,a;
+    byte r, g, b,a;
 
-    Color(int r = 255, int g = 255, int b = 255,int a=255) : r(r), g(g), b(b) ,a(a){}
+    Color(byte r = 255, byte g = 255, byte b = 255, byte a=255) : r(r), g(g), b(b) ,a(a){}
 
     // 线性插值
     static Color Lerp(const Color &start, const Color &end, float t);
